shell: add initial pmix plugin
Problem: a plugin offering pmix service is required to
bootstrap openmpi versions 5 and beyond.

Create an initial plugin that does the absolute minimum
AFAIK to be able to do some basic stuff on a single node/shell,
for example, execute a barrier.

Much will be filled in with subsequent commits.