github: add initial github workflow
Problem: The flux-pmix repository has no automated CI.

Add a simple GitHub Actions workflow to validate commits and run
docker-run-checks.sh against pull requests.

Currently, the checks are only performed against a focal container
with the v5.0.x branch of OpenMPI installed. However, the action
is setup so it will be easy to expand the matrix of builds to other
ompi versions and/or distros in the future.