plugin: set pmix.srvr.tmpdir
Problem: the pmix server creates its db files directly
in /tmp.  This may prove to be a cleanup nuisance.

Set pmix.srvr.tmpdir to FLUX_JOB_TMPDIR before
PMIx_server_init() is called, to inform the server where
they should be created.

Add a test to confirm that this works.