plugin: add codec for various data structures
Problem: we need to encode/decode the data structures used
in pmix callbacks for transport over interthread messages
and flux RPCs.

Add encode/decode functions for the common ones that are
likely to be needed.  Some, like the pmix_value_t, are not
completely implemented but should be sufficient for this
purpose.

Add unit test.