plugin: add direct_modex callback trace/warning
Problem: it is not clear yet whether the direct_modex server
callback is actually required.  It is complex enough that we
don't want to implement it if not needed;  however, it may be
hard to debug failures that could be attributed to its absence.

Implement the callback as a stub that always returns
PMIX_ERR_NOT_IMPLEMENTED.  Log each invocation with shell_warn().
In the future, if ompi is failing, the warning may be a useful
clue, and a suggestion to run with -overbose=2 to obtain the
dmodex_upcall interthread message payload trace and other
context may be good follow-up.