plugin: add interthread class
Problem: pmix callbacks are invoked in the pmix server thread
context, which is challenging to synchronize with the flux
reactor thread in the shell.

Introduce a class that implements a back-to-back shmem
message queue.  At the pmix server end, there is a send
function with a json_pack style message interface.  At the
shell end there is a callback registry by topic string.
The callbacks receive exactly one message, similar to flux
message handlers.

The messages are flux_msg_t, hardwired to always be request
type with a json payload.

Note that the message queue is bidirectional, but we only
use the server-to-shell direction.

The message queue is implemented with back to back shmem://
connectors, as demonstrated in the flux-core backtoback.c test.
This mechanism does have the disadvantage of using czmq, so
behind the scenes there is a zsys initialization that installs
an atexit handler, and may set up signal handlers for SIGINT
and SIGTERM.