plugin: add abort callback
Problem: we have not registered an abort callback, to be called
when a process calls PMIx_Abort().

Add callback, which calls shell_die() with the status and message
provided by the user, as we do in the built-in pmi plugin.

Note: this causes the shell to exit without finalizing plugins,
so some noise is generated by the czmq atexit handler because of
the unclosed interthread sockets.  Harmless but slightly annoying.
Leave it be for now.