plugin: add error notification callback
Problem: openpmix wants to notify flux when there is an internal
error using an error callback, but we don't have one.

The callback is part of a broader event notification system.
Register an event handler and print a warning if notification
is received.

N.B. users can also generate these events.