testsuite: fix uninitialized param in barrier test
Problem: barrier test fails on fedora34 with BAD-PARAM,
and sometimes segfaults.

If the --procs argument is not supplied to the barrier test program,
the 'procs' array argument is uninitialized.  This causes PMIx_Fence()
to fail on fedora34 (despite nprocs being set to zero).  In addition,
free() is later passed the uninitialized pointer, causing a segfault.

Initialize the procs array to NULL.