plugin: don't set pmix.nsdir
Problem: ompi occasionaly emits a dire warning about
an unlink failure when running mpi_hello on n2p4:
--------------------------------------------------------------------------
A system call failed during shared memory initialization that should
not have.  It is likely that your MPI job will now either abort or
experience performance degradation.

  Local host:  system76-pc0
  System call: unlink(2) /tmp/flux-OOSl1r/jobtmp-0-fQvfQmm/sm_segment.system76-pc0.5588.f5550000.1
  Error:       No such file or directory (errno 2)
--------------------------------------------------------------------------

I think this happens: In ompi/runtime/ompi_rte.c::ompi_rte_init(),
opal_process_info.job_session_dir is set to the value of pmix.nsdir,
if available, otherwise a new directory is created.
Then in ompi/runtime/ompi_rte.c::ompi_rte_finalize(),
opal_process_info.job_session_dir is recursively removed with the
comment /* cleanup the session directory we created */ without
first checking whether it came from pmix.nsdir.

pmix.nsdir is supposed to be set to a subdirectory of pmix.tmpdir,
but we set it to the same value as pmix.tmpdir, and pmix.tmpdir
contains shared memory segments.  So probably ompi is removing the
job_session_dir in a race with unlinking the shared memory segments.
If the job_session_dir is removed first, we get the above complaint.

Stop setting pmix.nsdir.  If we set it in the future, it should be
set to a subdirectory of pmix.tmpdir.

Reverse the pmix.nsdir test in t0002-basic.t (must NOT be set).

Add a regression test to t1000-ompi-basic.t.

Fixes #26