testsuite: add osu-micro-benchmarks-5.8
Problem: we should comprehensively exercise openmpi to
be sure we've not missed some bootstrap related setup
in the pmix shell plugin.

Pull in the osu-micro-benchmarks v. 5.8 from the tarball
posted here: https://mvapich.cse.ohio-state.edu/benchmarks/,
removing automake-related files and source not directly
related to the mpi tests.

A Makefile.am will be added in a subsequent commit.