plugin: use strlcpy()
Problem: strncpy() could leave a string unterminated if not
zero-initialized, and the keys for the info[2] array in main.c
not zero-initalized.

Switch strlcpy() wherever strncpy() is used.

Use sizeof() where appropriate instead of pmix "MAX" #defines,
since that removes all ambiguity about whether the #define includes
space for the NULL or not (it does not, but I have to keep checking).