plugin: set PMIX_PROC_INFO_ARRAY for local ranks
Problem: pmix.nodeid tests fail with openpmix 4.2.

Apparently we need to populate the "procs" info with an array
for each rank in the job.  Each array contains the global rank,
the local rank, and the node rank and is stored in the main info
array under the same key.

Note that the "node rank" is just set to the local rank.  What it
wants is a local rank that includes all jobs sharing the node,
but the shell doesn't have this information.

Fixes #59