testsuite: add debug to barrier test
Problem: the barrier test occasionally fails in CI with

expecting success:
  	run_timeout 30 flux mini run -N1 -n2 \
  		${BARRIER} --timeout=2

  fkdhs3d.0: completed PMIx_Init.
  fkdhs3d.1: PMIx_Fence: TIMEOUT
  fkdhs3d.0: there are 2 tasks
  fkdhs3d.0: completed barrier in 0.003s.
  fkdhs3d.0: completed PMIx_Finalize.
  0.106s: flux-shell[0]:  WARN: pmix: unknown optional fence attr: pmix.timeout
  0.106s: flux-shell[0]:  WARN: pmix: unknown optional fence attr: pmix.sorted.parr
  flux-job: task(s) exited with exit code 1
  not ok 3 - 1n2p barrier tolerates pmix.timeout=2

Add some debug to barrier.c so we can see the value of the timeout
option that is being passed in with the fence call on the client side.

Also run the test with -overbose=2 so we get server side info also.