build: allow build without openmpi
Problem: ompi is only needed for a subset of the test suite,
so it shouldn't prevent the package from being built.

Change configure logic so that ompi is optional.

If MPI is not configured, don't build the MPI-specific test
executables, except mpi_version which is used by sharness tests
to check if MPI is available.