upmi: add universal pmi client plugin
Problem: pmix client support (for bootstrapping Flux in an
environment where pmix service is provided by a foreign resource
manager) is conditionally compiled into flux-core.

It may not be convenient for the end user to rebuild flux-core,
for example if provided by a package manager.

Create a Flux upmi plugin for pmix that is provided by the flux-pmix
package.  This should let pmix support evolve independently of
flux-core.