testsuite: ensure pmix plugin is loaded
Problem: the pmix plugin is loaded in test but is no longer
active.

Update the test suite initrc.lua to set -o pmi=pmix if the
option is not manually set.

Drop -opmi=off from various places in the test suite.  The intent
was to disable simple pmi, which is now disabled by default.  Leaving
-opmi=off would also disable pmix.