testsuite: put flux bin before ompi bin
Problem: the test suite pushed the ompi bin directory in front
of the flux bin directory, but when ompi is installed as a system
package, this places /usr/bin in front of a possibly side installed
flux-core path.

This was noted to be the case on LLNL's lassen system.

Place the flux path in front of the ompi path when setting
up sharness test paths.

This was first noted in #85.