build: add configure --with-pmix[=PREFIX]
Problem: some versions of pmix don't have a .pc file, and
some packaged versions have a broken one.

Add a configure --with-pmix[=PREFIX].  If this is specified
without a PREFIX, default system paths are assumed.  If PREFIX
is specified, then PMIX_CFLAGS and PMIX_LIBS are set based on that.

No checks are performed to ensure PREFIX refers to a working
pmix install of the minimum version.  This is intended to be an
override mechanism for exceptional situations.  The default
pkg-config method is the preferred one, when it works.