plugin: use a message queue for interthread messages
Problem: the shell plugin uses the flux shmem:// connector
for interthread communication, but this now requires a 0MQ
context to be passed in.

Rather than making libzmq a requirement for flux-pmix and
creating a zeromq context to be passed in, just convert
the interthread channel to use a 'struct flux_msglist'
protected by mutexes.  The msglist class has a built in
pollfd/pollevents interface for reactor integration.

This also avoids potential issues with managing a zeromq context
in the flux shell, such as hanging when sockets are not properly
destroyed, and libzmq.so's library init taking a long time for
the crypto self-test on FIPS-enabled systems.

Fixes #89