plugin: set hwloc.xmlfile shell option
Problem: flux-pmix does not yet share hwloc with MPI, which
forces MPI to go looking for it, sometimes at great cost
to performance.

For the time being, tell flux-core to share a hwloc xml
file by setting the -o hwloc.xmlfile shell option.

Later we will want to fix #31 properly and drop this.