interoperate with older flux-core releases
Problem: the ABI changes in flux-core 0.70 to raw payload
accessors actually need not impact this project since all
interthread payloads are NUL terminated strings.

The last parameter of flux_msg_get_payload() is changing from an
int pointer to a size_t pointer.  Just pass a NULL and avoid
the entire issue as we don't need the size.