simplify interthread channel implementation
Problem: the interthread channel will be harder to maintain than
necessary due to its use of the obscure flux_msglist pollfd/pollevents
interface, an fd watcher, and a pthread mutex.

Now that we have the interthread:// connector in flux-core,
reimplement the interthread channel with back to back flux_t handles
and a message watcher.