configure: allow --without-openmpi to disable Open MPI
Problem: There is no way to disable MPI when there is an OpenMPI
installation that can be found in the default pkg-config path.
However, since MPI is only used for testing, disabling it for
the build is a valid use case.

Add a --without-openmpi option to configure that allows MPI to
be disabled, even when OpenMPI would be found by default.

Fixes #114